pyproject.toml
[project]
name = "radio-browser-mcp"
version = "0.1.0"
description = "Radio Browser and Playback MCP Server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp",
    "requests>=2.28.0",
    "python-vlc"
]

[project.scripts]
radio-browser-mcp = "server:mcp.run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []
